requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
openai==2.7.1
psycopg2-binary==2.9.11
requests==2.28.2
Werkzeug==3.1.3
